requirements.txt
pytz==2013.7
requests==2.18.1
beautifulsoup4==4.6.0
feedparser==5.2.1
pytidylib==0.3.2
